Backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
pymysql==1.1.0
mysql-connector-python==8.3.0
torch==2.2.0+cu118
torchvision==0.17.0+cu118
ultralytics==8.1.0
opencv-python==4.9.0.80
numpy<2
scikit-learn==1.4.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
groq==0.4.1

--extra-index-url https://download.pytorch.org/whl/cu118
